one button was added to the grid but there should be three buttons
