+++ bin/.gitignore
@@ -1,2 +1,8 @@
 /Control/
+/FileManagement/
+/GUI/
 /Model/
+/com/
+/infoWindow.fxml
+/net/
+/theWindow.fxml
