+++ src/Control/startProjectControl.java
@@ -1,25 +1,30 @@
 package Control;
 
 import java.io.IOException;
+import java.net.URL;
+import java.util.ResourceBundle;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import Model.Project;
 import Model.ProjectManager;
+import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
+import javafx.fxml.Initializable;
 import javafx.scene.Node;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
+import javafx.scene.control.TableColumn;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.GridPane;
 import javafx.stage.Stage;
 
 
-public class startProjectControl {
+public class startProjectControl{
 	//private WindowControl main;
     @FXML
     private Button cancel;
@@ -42,6 +47,7 @@ public class startProjectControl {
 	private ProjectManager pm ;
 	private WindowControl main;
 	Project project;
+	ObservableList<Project> data;
 
     @FXML
     void handelCancelButtonAction() {
@@ -55,8 +61,7 @@ public class startProjectControl {
 		if (nameValidation(((TextField) input.get(0)).getText())) {
 			projectName += ((TextField) input.get(0)).getText();
 			errorLabel.setVisible(false);
-			pm.addProject(project = new Project(projectName));
-			add_new_project_to_grid(create_button(projectName));
+			pm.addProject(new Project(projectName));
 			((Stage)(finish.getScene().getWindow())).close();
 		}else {
 			errorLabel.setVisible(true);
@@ -69,40 +74,24 @@ public class startProjectControl {
 		return m.matches();
     }
     
- 
+   
     
-public void add_new_project_to_grid(Button button) {
-	ObservableList<Node> index = main.gp.getChildren();
-	int i = 0;
-	if(index.get(i) == null) {
-		main.add_project_to_Grid(button);
-	}else {
-		while(index.get(i) != null) {
-			i++;
-		}
-		main.add_project_to_Grid(button);
-	}
-}
     
-   public Button create_button(String name) {	
-	
-		Button button = new Button(name);
-		System.out.println(button.getText());
-    	button.setOnAction(e->{
-    		 try {
-    	            FXMLLoader loader= new FXMLLoader(getClass().getResource("/NewProject.fxml"));
-    	            AnchorPane Ap =  loader.load();
-    	            Stage stage = new Stage();
-    	            Scene window = new Scene(Ap);
-    	            stage.setScene(window);
-    	            stage.show();
-    	        }
-    	        catch (IOException e1) {
-    	            e1.printStackTrace();
-    	       }
-    	    
-    	});
-    	return button;
-    }   
+    
+    
+//public void add_new_project_to_grid(Button button) {
+//	ObservableList<Node> index = main.gp.getChildren();
+//	int i = 0;
+//	if(index.get(i) == null) {
+//		//main.add_project_to_Grid(button);
+//	}else {
+//		while(index.get(i) != null) {
+//			i++;
+//		}
+//		//main.add_project_to_Grid(button);
+//	}
+//}
+    
+  
     
 }
